+++ tests/test_new_command.rs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::path::Path;
+
 use crate::common::TestEnvironment;
 
 pub mod common;
@@ -25,8 +27,7 @@ fn test_new() {
     test_env.jj_cmd_success(&repo_path, &["describe", "-m", "add a file"]);
     test_env.jj_cmd_success(&repo_path, &["new", "-m", "a new commit"]);
 
-    let stdout = test_env.jj_cmd_success(&repo_path, &["log", "-T", "commit_id \" \" description"]);
-    insta::assert_snapshot!(stdout, @r###"
+    insta::assert_snapshot!(get_log_output(&test_env, &repo_path), @r###"
     @ 88436dbcdbedc2b8a6ebd0687981906d09ccc68f a new commit
     o 51e9c5819117991e4a6dc5a4a744283fc74f0746 add a file
     o 0000000000000000000000000000000000000000 (no description set)
@@ -34,12 +35,43 @@ fn test_new() {
 
     // Start a new change off of a specific commit (the root commit in this case).
     test_env.jj_cmd_success(&repo_path, &["new", "-m", "off of root", "root"]);
-    let stdout = test_env.jj_cmd_success(&repo_path, &["log", "-T", "commit_id \" \" description"]);
-    insta::assert_snapshot!(stdout, @r###"
+    insta::assert_snapshot!(get_log_output(&test_env, &repo_path), @r###"
     @ d8c0a3e1570f1f5b08113a3427b3160900c3d48e off of root
     | o 88436dbcdbedc2b8a6ebd0687981906d09ccc68f a new commit
     | o 51e9c5819117991e4a6dc5a4a744283fc74f0746 add a file
     |/  
     o 0000000000000000000000000000000000000000 (no description set)
     "###);
+}
+
+#[test]
+fn test_new_merge() {
+    let test_env = TestEnvironment::default();
+    test_env.jj_cmd_success(test_env.env_root(), &["init", "repo", "--git"]);
+    let repo_path = test_env.env_root().join("repo");
+
+    test_env.jj_cmd_success(&repo_path, &["branch", "create", "main"]);
+    test_env.jj_cmd_success(&repo_path, &["describe", "-m", "add file1"]);
+    std::fs::write(repo_path.join("file1"), "a").unwrap();
+    test_env.jj_cmd_success(&repo_path, &["new", "root", "-m", "add file2"]);
+    std::fs::write(repo_path.join("file2"), "b").unwrap();
+
+    // Create a merge commit
+    test_env.jj_cmd_success(&repo_path, &["new", "main", "@"]);
+    insta::assert_snapshot!(get_log_output(&test_env, &repo_path), @r###"
+    @   5b37ef8ee8cd934dfe1e70adff66cd0679f5a573 (no description set)
+    |\  
+    o | 99814c62bec5c13d2053435b3d6bbeb1900cb57e add file2
+    | o fe37af248a068697c6dcd7ebd17f5aac2205e7cb add file1
+    |/  
+    o 0000000000000000000000000000000000000000 (no description set)
+    "###);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["print", "file1"]);
+    insta::assert_snapshot!(stdout, @"a");
+    let stdout = test_env.jj_cmd_success(&repo_path, &["print", "file2"]);
+    insta::assert_snapshot!(stdout, @"b");
+}
+
+fn get_log_output(test_env: &TestEnvironment, repo_path: &Path) -> String {
+    test_env.jj_cmd_success(repo_path, &["log", "-T", "commit_id \" \" description"])
 }