+++ CHANGELOG.md
@@ -138,6 +138,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * `jj merge` now outputs the id of the newly created commit.
 
+* `jj new` can now be used for creating merge commits. If you pass more than
+  one argument to it, the new commit will have all of them as parents.
+
 ### Fixed bugs
 
 * When rebasing a conflict where one side modified a file and the other side